Moved camera related logic to renamed class LeapCameraControl
